changed deployment issues

+++ app/run.sh
@@ -14,11 +14,16 @@ mysql -u root -e "CREATE DATABASE IF NOT EXISTS wikipanime;"
 mysql -u root -e "CREATE USER IF NOT EXISTS 'yuwkaa'@'localhost' IDENTIFIED BY 'fake_db_password';"
 mysql -u root -e "GRANT ALL PRIVILEGES ON wikipanime.* TO 'yuwkaa'@'localhost';"
 
-echo "Loading data from init.sql..."
-mysql -u root wikipanime < /tmp/init.sql
+# Kiem tra xem file init.sql co ton tai khong truoc khi nap
+if [ -f "/tmp/init.sql" ]; then
+    echo "Loading data from init.sql..."
+    mysql -u root wikipanime < /tmp/init.sql
+    
+    echo "Removing init.sql..."
+    rm -f /tmp/init.sql
+else
+    echo "init.sql not found. Database is already initialized, skipping..."
+fi
 
-echo "Removing init.sql..."
-rm -f /tmp/init.sql
-
-echo "Starting Flask as user ctf..."
-su ctf -c "python3 app.py"
+echo "Starting Flask as user nguoiyeuhayaseyuuka..."
+su nguoiyeuhayaseyuuka -c "python3 app.py"
